\name{badClassStop}
\alias{badClassStop}
\title{Stop Function Execution on Wrong Class }
\usage{
badClassStop(x, class)
}
\description{Calls \code{stop} if \code{x} is not a \code{class}}
\arguments{
  \item{x}{ an \R object }
  \item{class}{ character string }
}
\author{ Jeff Hallman }
\keyword{ classes }
\keyword{ programming }
