\name{stripBlanks}
\alias{stripBlanks}
\title{Strip Blanks}
\description{Strips leading and trailing blanks from strings}
\usage{
stripBlanks(strings)
}
\arguments{
  \item{strings}{character vector}
}
\value{An object like \code{strings} with no leading or trailing blanks.}
\author{ Jeff Hallman }
\seealso{ \code{\link{blanks}}, \code{\link{gsub}} }
\keyword{character}
